# docs/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Vaeloom Backend
  version: 0.2.0
paths:
...
  /api/v1/provider-keys:
    get:
      tags:
        - provider-keys
      summary: List Provider Keys
      operationId: list_provider_keys_api_v1_provider_keys_get
      parameters:
        - name: workspace_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderKeyListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - provider-keys
      summary: Create Provider Key
      operationId: create_provider_key_api_v1_provider_keys_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderKeyCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderKeyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/provider-keys/effective:
    get:
      tags:
        - provider-keys
      summary: Get Effective Key
      operationId: get_effective_key_api_v1_provider_keys_effective_get
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            title: Provider
        - name: workspace_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectiveKeyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/provider-keys/{key_id}:
    delete:
      tags:
        - provider-keys
      summary: Delete Provider Key
      operationId: delete_provider_key_api_v1_provider_keys__key_id__delete
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Key Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - provider-keys
      summary: Update Provider Key
      operationId: update_provider_key_api_v1_provider_keys__key_id__patch
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Key Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderKeyUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderKeyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/provider-keys/{key_id}/validate:
    post:
      tags:
        - provider-keys
      summary: Validate Provider Key
      operationId: validate_provider_key_api_v1_provider_keys__key_id__validate_post
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Key Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateKeyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    EffectiveKeyResponse:
      properties:
        provider:
          type: string
          title: Provider
        has_custom_key:
          type: boolean
          title: Has Custom Key
        source:
          type: string
          title: Source
        key_hint:
          anyOf:
            - type: string
            - type: 'null'
          title: Key Hint
        is_valid:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Is Valid
        is_active:
          type: boolean
          title: Is Active
          default: true
        last_validated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Last Validated At
        fallback_provider:
          anyOf:
            - type: string
            - type: 'null'
          title: Fallback Provider
      type: object
      required:
        - provider
        - has_custom_key
        - source
      title: EffectiveKeyResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProviderKeyCreate:
      properties:
        provider:
          type: string
          title: Provider
          description: 'LLM provider key id: openai | anthropic | google | mistral
            | cohere | azure | openrouter | groq | ollama | custom'
        api_key:
          type: string
          maxLength: 500
          minLength: 8
          title: Api Key
          description: Raw provider API key (never returned plaintext)
        workspace_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Workspace Id
          description: If null => user-global key; if set => workspace-scoped override
      type: object
      required:
        - provider
        - api_key
      title: ProviderKeyCreate
    ProviderKeyListResponse:
      properties:
        keys:
          items:
            $ref: '#/components/schemas/ProviderKeyResponse'
          type: array
          title: Keys
        total:
          type: integer
          title: Total
      type: object
      required:
        - keys
        - total
      title: ProviderKeyListResponse
    ProviderKeyResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        provider:
          type: string
          title: Provider
        key_hint:
          type: string
          title: Key Hint
        key_prefix:
          type: string
          title: Key Prefix
        is_active:
          type: boolean
          title: Is Active
        is_valid:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Is Valid
        last_validated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Last Validated At
        last_used_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Last Used At
        validation_error:
          anyOf:
            - type: string
            - type: 'null'
          title: Validation Error
        workspace_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Workspace Id
        user_id:
          type: string
          format: uuid
          title: User Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
        - id
        - provider
        - key_hint
        - key_prefix
        - is_active
        - user_id
        - created_at
        - updated_at
      title: ProviderKeyResponse
    ProviderKeyUpdate:
      properties:
        api_key:
          anyOf:
            - type: string
              maxLength: 500
              minLength: 8
            - type: 'null'
          title: Api Key
        is_active:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Is Active
      type: object
      title: ProviderKeyUpdate
...
    ValidateKeyResponse:
      properties:
        is_valid:
          type: boolean
          title: Is Valid
        provider:
          type: string
          title: Provider
        message:
          type: string
          title: Message
        latency_ms:
          anyOf:
            - type: integer
            - type: 'null'
          title: Latency Ms
      type: object
      required:
        - is_valid
        - provider
        - message
      title: ValidateKeyResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...